+++ extensions/reactive-oracle-client/deployment/src/test/java/io/quarkus/reactive/oracle/client/LocalhostOraclePoolCreator.java
@@ -2,7 +2,6 @@ package io.quarkus.reactive.oracle.client;
 
 import javax.inject.Singleton;
 
-import io.quarkus.reactive.oracle.client.runtime.OraclePoolCreator;
 import io.vertx.oracleclient.OraclePool;
 
 @Singleton
