+++ extensions/reactive-oracle-client/deployment/src/test/java/io/quarkus/reactive/oracle/client/MultipleOraclePoolCreatorsForSameDatasourceTest.java
@@ -8,7 +8,6 @@ import javax.inject.Singleton;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-import io.quarkus.reactive.oracle.client.runtime.OraclePoolCreator;
 import io.quarkus.test.QuarkusUnitTest;
 import io.vertx.oracleclient.OraclePool;
